requirements.txt
guicorn
flask
tensorflow
numpy
pillow
flaswagger
flask_cors
sklearn